+++ CHANGELOG.md
@@ -1,1 +1,3 @@
 - Support 1.21.10
+- #18 Replace `MinecraftServer.getResourcePackProperties` instead of `SendResourcePackTask`'s constructor for
+better compatibility with other mods